refactor: move background components to separate folder under background
